// src/app/customer-profile/model/measurement.ts
export class Measurement{
  id: number;
  coatLength: number;
  coatChest: number;
  coatWaist: number;
  coatHip: number
  coatShoulder: number;
  coatSleeveLength: number;
  coatHalfBack: number;
  coatNeck: number;
  shirtLength: number;
  shirtChest: number;
  shirtWaist: number;
  shirtHip: number;
  shirtShoulder: number;
  shirtSleeveLength: number;
  shirtNeck: number;
  pantLength: number;
  pantWaist: number;
  pantHip: number;
  pantThigh: number;
  pantKnee: number;
  pantBottom: number;
  pantCrouch: number;
  overCoatLength: number;
  waistCoatLength: number;
  dauraLength: number;
  sherwaniLength: number;


}